requirements.txt
opencv-python~=4.10.0.84
numpy~=2.1.3
torch~=2.5.1+cu118
torchvision~=0.20.1+cu118
pillow~=11.0.0
matplotlib~=3.9.2
pandas~=2.2.3
ultralytics~=8.3.48